backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pyodbc==5.0.1
python-dotenv==1.0.0
google-generativeai==0.3.2

